+++ ICD.Common.Logging/Fusion/FusionErrorEventArgs.cs
@@ -1,7 +1,7 @@
 using System;
 using ICD.Common.Properties;
 
-namespace ICD.Common.Logging.Console.Fusion
+namespace ICD.Common.Logging.Fusion
 {
 	public sealed class FusionErrorEventArgs : EventArgs
 	{