+++ ICD.Common.Logging/LoggingCore.cs
@@ -1,13 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
-using ICD.Common.Logging.Console.Loggers;
+using ICD.Common.Logging.Loggers;
 using ICD.Common.Properties;
 using ICD.Common.Utils;
 using ICD.Common.Utils.Collections;
 using ICD.Common.Utils.Extensions;
 using ICD.Common.Utils.Services.Logging;
 
-namespace ICD.Common.Logging.Console
+namespace ICD.Common.Logging
 {
 	/// <summary>
 	/// Core of the ELogging functionality.