+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+ - Renamed ICD.Common.Logging.Console project to ICD.Common.Logging
+
 ## [5.1.0] - 2019-04-16
 ### Added
  - Added FileLogger